+++ Repository/BaseRepository/IGenericRepository.cs
@@ -8,9 +8,9 @@ namespace Repository
     public interface IGenericRepository<T> where T : class
     {
         T Save(T entity);
-        T GetById(long id);
+        T GetById(int id);
         T Update(T entity);
-        bool DeleteById(long id);
+        bool DeleteById(int id);
         bool BulkInsert(List<T> entityList);
         bool BulkDelete(List<T> entityList);
         IQueryable<T> GetAll();
